Code climate fix
