+++ app/src/main/java/org/ole/planet/myplanet/ui/course/CourseFragment.java
@@ -79,9 +79,7 @@ public class CourseFragment extends BaseRecyclerFragment<RealmMyCourse> implemen
             KeyboardUtils.hideSoftKeyboard(getActivity());
         });
         // setSearchListener();
-        btnRemove.setOnClickListener(V -> {
-            deleteSelected(true);
-        });
+        btnRemove.setOnClickListener(V -> { deleteSelected(true); });
         getView().findViewById(R.id.btn_collections).setOnClickListener(view -> {
             CollectionsFragment f = CollectionsFragment.getInstance(searchTags, "courses");
             f.setListener(this);
@@ -92,6 +90,10 @@ public class CourseFragment extends BaseRecyclerFragment<RealmMyCourse> implemen
         KeyboardUtils.setupUI(getView().findViewById(R.id.my_course_parent_layout), getActivity());
         changeButtonStatus();
         if(!isMyCourseLib) tvFragmentInfo.setText("Our Courses");
+        additionalSetup();
+    }
+
+    public void additionalSetup(){
         orderByDate = getView().findViewById(R.id.order_by_date_button);
         orderByTitle = getView().findViewById(R.id.order_by_title_button);
         orderByDate.setOnClickListener(view -> adapterCourses.setCourseList(getList(RealmMyCourse.class,"createdDate")));
